+++ lib/interactable_svg/src/parser.dart
@@ -30,7 +30,8 @@ class Parser {
       final region = Region(
           id: regionData.group(1)!,
           name: regionData.group(2)!,
-          path: parseSvgPath(regionData.group(3)!));
+          className: regionData.group(3),
+          path: parseSvgPath(regionData.group(4)!));
 
       sizeController.addBounds(region.path.getBounds());
       regionList.add(region);
@@ -45,12 +46,12 @@ class Parser {
 
     final regExp = RegExp(Constants.mapRegexp,
         multiLine: true, caseSensitive: false, dotAll: false);
-
     regExp.allMatches(svgMain).forEach((regionData) {
       final region = Region(
           id: regionData.group(1)!,
           name: regionData.group(2)!,
-          path: parseSvgPath(regionData.group(3)!));
+          className: regionData.group(3),
+          path: parseSvgPath(regionData.group(4)!));
 
       sizeController.addBounds(region.path.getBounds());
       regionList.add(region);
@@ -73,7 +74,8 @@ class Parser {
       final region = Region(
           id: regionData.group(1)!,
           name: regionData.group(2)!,
-          path: parseSvgPath(regionData.group(3)!));
+          className: regionData.group(3),
+          path: parseSvgPath(regionData.group(4)!));
 
       sizeController.addBounds(region.path.getBounds());
       regionList.add(region);